+++ plugins/actix-web/src/lib.rs
@@ -1,17 +1,25 @@
+#![cfg(any(feature = "actix2", feature = "actix3"))]
+
 #[cfg(feature = "actix2")]
 extern crate actix_web2 as actix_web;
 #[cfg(feature = "actix3")]
 extern crate actix_web3 as actix_web;
 
+#[cfg(feature = "swagger-ui")]
+use include_dir::{include_dir, Dir};
+
 pub mod web;
 
 pub use self::web::{Resource, Route, Scope};
 pub use paperclip_macros::{
-    api_v2_errors, api_v2_operation, delete, get, post, put, Apiv2Schema, Apiv2Security,
+    api_v2_errors, api_v2_errors_overlay, api_v2_operation, delete, get, post, put, Apiv2Schema,
+    Apiv2Security,
 };
 
 use self::web::{RouteWrapper, ServiceConfig};
 use actix_service::ServiceFactory;
+#[cfg(feature = "swagger-ui")]
+use actix_web::web::HttpRequest;
 use actix_web::{
     dev::{HttpServiceFactory, MessageBody, ServiceRequest, ServiceResponse, Transform},
     web::HttpResponse,
@@ -29,8 +37,15 @@ use std::{collections::BTreeMap, fmt::Debug, future::Future, sync::Arc};
 /// Wrapper for [`actix_web::App`](https://docs.rs/actix-web/*/actix_web/struct.App.html).
 pub struct App<T, B> {
     spec: Arc<RwLock<DefaultApiRaw>>,
+    #[cfg(feature = "v3")]
+    spec_v3: Option<Arc<RwLock<openapiv3::OpenAPI>>>,
+    #[cfg(feature = "swagger-ui")]
+    spec_path: Option<String>,
     inner: Option<actix_web::App<T, B>>,
 }
+
+#[cfg(feature = "swagger-ui")]
+static SWAGGER_DIST: Dir = include_dir!("../swagger-ui/dist");
 
 /// Extension trait for actix-web applications.
 pub trait OpenApiExt<T, B> {
@@ -52,6 +67,10 @@ impl<T, B> OpenApiExt<T, B> for actix_web::App<T, B> {
     fn wrap_api(self) -> Self::Wrapper {
         App {
             spec: Arc::new(RwLock::new(DefaultApiRaw::default())),
+            #[cfg(feature = "v3")]
+            spec_v3: None,
+            #[cfg(feature = "swagger-ui")]
+            spec_path: None,
             inner: Some(self),
         }
     }
@@ -59,6 +78,10 @@ impl<T, B> OpenApiExt<T, B> for actix_web::App<T, B> {
     fn wrap_api_with_spec(self, spec: DefaultApiRaw) -> Self::Wrapper {
         App {
             spec: Arc::new(RwLock::new(spec)),
+            #[cfg(feature = "v3")]
+            spec_v3: None,
+            #[cfg(feature = "swagger-ui")]
+            spec_path: None,
             inner: Some(self),
         }
     }
@@ -223,6 +246,10 @@ where
     {
         App {
             spec: self.spec,
+            #[cfg(feature = "v3")]
+            spec_v3: self.spec_v3,
+            #[cfg(feature = "swagger-ui")]
+            spec_path: None,
             inner: self.inner.take().map(|a| a.wrap(mw)),
         }
     }
@@ -250,6 +277,10 @@ where
     {
         App {
             spec: self.spec,
+            #[cfg(feature = "v3")]
+            spec_v3: self.spec_v3,
+            #[cfg(feature = "swagger-ui")]
+            spec_path: None,
             inner: self.inner.take().map(|a| a.wrap_fn(mw)),
         }
     }
@@ -258,10 +289,37 @@ where
     /// recorded by the wrapper and serves them in the given path
     /// as a JSON.
     pub fn with_json_spec_at(mut self, path: &str) -> Self {
+        #[cfg(feature = "swagger-ui")]
+        {
+            self.spec_path = Some(path.to_owned());
+        }
+
         self.inner = self.inner.take().map(|a| {
             a.service(
                 actix_web::web::resource(path)
                     .route(actix_web::web::get().to(SpecHandler(self.spec.clone()))),
+            )
+        });
+        self
+    }
+
+    #[cfg(feature = "v3")]
+    /// Converts the generated v2 specification to v3 and then
+    /// mounts the v3 specification for all operations and definitions
+    /// recorded by the wrapper and serves them in the given path
+    /// as a JSON.
+    pub fn with_json_spec_v3_at(mut self, path: &str) -> Self {
+        let spec_v3 = if let Some(spec_v3) = &self.spec_v3 {
+            spec_v3.clone()
+        } else {
+            let spec_v3 = Arc::new(RwLock::new(openapiv3::OpenAPI::default()));
+            self.spec_v3 = Some(spec_v3.clone());
+            spec_v3
+        };
+        self.inner = self.inner.take().map(|a| {
+            a.service(
+                actix_web::web::resource(path)
+                    .route(actix_web::web::get().to(SpecHandlerV3(spec_v3.clone()))),
             )
         });
         self
@@ -281,9 +339,89 @@ where
         call(self, spec)
     }
 
+    #[cfg(feature = "v3")]
+    /// Calls the given function with `App` and JSON `Value` representing your API
+    /// v2 specification **built until now** which is converted to v3.
+    ///
+    /// **NOTE:** Unlike `with_json_spec_at`, this only has the API spec built until
+    /// this function call. Any route handler added after this call won't affect the
+    /// spec. So, it's important to call this function after adding all route handlers.
+    pub fn with_raw_json_spec_v3<F>(self, mut call: F) -> Self
+    where
+        F: FnMut(Self, serde_json::Value) -> Self,
+    {
+        let v3 = paperclip_core::v3::openapiv2_to_v3(self.spec.read().clone());
+        let spec = serde_json::to_value(v3).expect("generating json spec");
+        call(self, spec)
+    }
+
+    /// Exposes the previously built JSON specification with Swagger UI at the given path
+    ///
+    /// **NOTE:** you **MUST** call with_json_spec_at before calling this function
+    #[cfg(feature = "swagger-ui")]
+    pub fn with_swagger_ui_at(mut self, path: &str) -> Self {
+        let spec_path = self.spec_path.clone().expect(
+            "Specification not set, be sure to call `with_json_spec_at` before this function",
+        );
+
+        let path: String = path.into();
+        // Grab any file request from the documentation UI path and fetch it from SWAGGER_DIST
+        // E.g: js, html, svg and etc.
+        let regex_path = format!("{}/{{filename:.*}}", path);
+
+        self.inner = self.inner.take().map(|a| {
+            a.service(
+                actix_web::web::resource([regex_path.to_owned(), path.clone()]).route(
+                    actix_web::web::get().to(move |request: HttpRequest| {
+                        let filename = request.match_info().query("filename");
+                        if filename.is_empty() && request.query_string().is_empty() {
+                            let redirect_url = format!("{}/index.html?url={}", path, spec_path);
+                            HttpResponse::PermanentRedirect()
+                                .header("Location", redirect_url)
+                                .finish()
+                        } else {
+                            HttpResponse::Ok().body(
+                                SWAGGER_DIST
+                                    .get_file(filename)
+                                    .unwrap()
+                                    .contents_utf8()
+                                    .unwrap(),
+                            )
+                        }
+                    }),
+                ),
+            )
+        });
+        self
+    }
+
     /// Builds and returns the `actix_web::App`.
     pub fn build(self) -> actix_web::App<T, B> {
+        #[cfg(feature = "v3")]
+        self.spec_v3.clone().map(|v3| {
+            let mut v3 = v3.write();
+            *v3 = paperclip_core::v3::openapiv2_to_v3(self.spec.read().clone());
+        });
         self.inner.expect("missing app?")
+    }
+
+    /// Trim's the Api base path from the start of all method paths.
+    /// **NOTE:** much like `with_raw_json_spec` this only has the API spec built until
+    /// this function call. Any route handler added after this call won't have the base path trimmed.
+    /// So, it's important to call this function after adding all route handlers.
+    pub fn trim_base_path(self) -> Self {
+        {
+            let mut spec = self.spec.write();
+            let base_path = spec.base_path.clone().unwrap_or_default();
+            spec.paths = spec.paths.iter().fold(BTreeMap::new(), |mut i, (k, v)| {
+                i.insert(
+                    k.trim_start_matches(base_path.as_str()).to_string(),
+                    v.clone(),
+                );
+                i
+            });
+        }
+        self
     }
 
     /// Updates the underlying spec with definitions and operations from the given factory.
@@ -311,6 +449,19 @@ struct SpecHandler(Arc<RwLock<DefaultApiRaw>>);
 
 impl actix_web::dev::Factory<(), Ready<Result<HttpResponse, Error>>, Result<HttpResponse, Error>>
     for SpecHandler
+{
+    fn call(&self, _: ()) -> Ready<Result<HttpResponse, Error>> {
+        fut_ok(HttpResponse::Ok().json(&*self.0.read()))
+    }
+}
+
+#[cfg(feature = "v3")]
+#[derive(Clone)]
+struct SpecHandlerV3(Arc<RwLock<openapiv3::OpenAPI>>);
+
+#[cfg(feature = "v3")]
+impl actix_web::dev::Factory<(), Ready<Result<HttpResponse, Error>>, Result<HttpResponse, Error>>
+    for SpecHandlerV3
 {
     fn call(&self, _: ()) -> Ready<Result<HttpResponse, Error>> {
         fut_ok(HttpResponse::Ok().json(&*self.0.read()))